+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.4.0 (IN PROGRESS)
 
 * Print hold slip at check in. Fixes UICHKIN-16.
+* Extract hardcoded labels. Fixes UICHKIN-14.
 
 ## [1.3.0](https://github.com/folio-org/ui-checkin/tree/v1.3.0) (2018-10-04)
 [Full Changelog](https://github.com/folio-org/ui-checkin/compare/v1.2.0...v1.3.0)