+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Handle check in for items that are In transit. Fixes UICHKIN-44.
 * Display destination service point for in transit items. Fixes UICHKIN-40.
 * Display modal when items go in transit. Fixes UICHKIN-41.
+* Fix link to item. Fixes UICHKIN-61.
 
 ## [1.3.0](https://github.com/folio-org/ui-checkin/tree/v1.3.0) (2018-10-04)
 [Full Changelog](https://github.com/folio-org/ui-checkin/compare/v1.2.0...v1.3.0)