+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change history for ui-checkin
 
+## 1.5.0 IN PROGRESS
+
 ## [1.4.0](https://github.com/folio-org/ui-checkin/tree/v1.4.0) (2018-12-13)
 [Full Changelog](https://github.com/folio-org/ui-checkin/compare/v1.3.0...v1.4.0)